+++ app/chess/page.tsx
@@ -1,6 +1,6 @@
 import { Poppins } from "next/font/google";
 
-import ChessPage from "../components/ChessPage";
+import ChessPage from "@/app/components/chess/ChessPage";
 
 const chessFont = Poppins({ weight: "400", subsets: ["latin"] });
