+++ app/components/chess/ChessContent.tsx
