+++ app/components/landing/MagicBento.tsx
