+++ app/components/shared/display/DecryptedText.tsx
