+++ app/components/shared/feedback/PageSkeletons.tsx
@@ -1,4 +1,4 @@
-import { Skeleton } from "@/components/ui/skeleton";
+import { Skeleton } from "@/app/components/shared/ui/skeleton";
 
 const shellContainerClassName =
   "mx-auto max-sm:w-[300px] max-xs:w-[230px] sm:w-[560px] md:w-[680px] lg:w-[910px] xl:w-[1160px] xxl:w-[1480px]";
