+++ app/components/shared/navigation/StaggeredMenu.jsx
